Scripts/GitHub/requirements.txt
# requirements.txt

requests
python-dotenv
